+++ packages/fighting-design/_hooks/use-message-work/index.ts
@@ -107,8 +107,6 @@ export const useMessageWork = (
     return styles
   })
 
-  // console.log(offsetStyle.value)
-
   /** 清除计时器 */
   const clearTimer = (): void => {
     if (!timeout) return