+++ packages/fighting-design/count-down/src/count-down.vue
@@ -72,8 +72,6 @@
       }
     }
 
-    console.log(format)
-
     return format
   }